+++ publication-file/src/main/java/no/unit/nva/publication/file/upload/config/MultipartUploadConfig.java
@@ -1,0 +1,11 @@
+package no.unit.nva.publication.file.upload.config;
+
+import nva.commons.core.Environment;
+
+public final class MultipartUploadConfig {
+
+    public static final String BUCKET_NAME = new Environment().readEnv("NVA_PERSISTED_STORAGE_BUCKET_NAME");
+
+    private MultipartUploadConfig() {
+    }
+}
