+++ publication-file/src/main/java/no/unit/nva/publication/file/upload/restmodel/CreateUploadResponseBody.java
@@ -1,0 +1,10 @@
+package no.unit.nva.publication.file.upload.restmodel;
+
+import com.amazonaws.services.s3.model.InitiateMultipartUploadResult;
+
+public record CreateUploadResponseBody(String uploadId, String key) {
+
+    public static CreateUploadResponseBody fromInitiateMultipartUploadResult(InitiateMultipartUploadResult result) {
+        return new CreateUploadResponseBody(result.getUploadId(), result.getKey());
+    }
+}
