+++ publication-file/src/main/java/no/unit/nva/publication/file/upload/restmodel/ListPartsResponseBody.java
@@ -1,0 +1,7 @@
+package no.unit.nva.publication.file.upload.restmodel;
+
+import java.util.List;
+
+public record ListPartsResponseBody(List<ListPartsElement> listParts) {
+
+}
