+++ publication-file/src/main/java/no/unit/nva/publication/file/upload/restmodel/PrepareUploadPartResponseBody.java
@@ -1,0 +1,7 @@
+package no.unit.nva.publication.file.upload.restmodel;
+
+import java.net.URL;
+
+public record PrepareUploadPartResponseBody(URL url) {
+
+}
