+++ publication-file/src/test/resources/CompleteRequestWithEmptyElement.json
