--- publication-rest/src/main/java/no/unit/nva/publication/uploadfile/restmodel/CompleteUploadPart.java
@@ -1,22 +1,0 @@
-package no.unit.nva.publication.uploadfile.restmodel;
-
-import static java.util.Objects.requireNonNull;
-import com.amazonaws.services.s3.model.PartETag;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record CompleteUploadPart(@JsonProperty("PartNumber") Integer partNumber, @JsonProperty("ETag") String etag) {
-
-    public static PartETag toPartETag(CompleteUploadPart completeUploadPart) {
-        return new PartETag(completeUploadPart.partNumber(), completeUploadPart.etag());
-    }
-
-    public boolean hasValue() {
-        try {
-            requireNonNull(etag);
-            requireNonNull(partNumber);
-        } catch (Exception e) {
-            return false;
-        }
-        return true;
-    }
-}
