--- publication-rest/src/main/java/no/unit/nva/publication/uploadfile/restmodel/ListPartsResponseBody.java
@@ -1,7 +1,0 @@
-package no.unit.nva.publication.uploadfile.restmodel;
-
-import java.util.List;
-
-public record ListPartsResponseBody(List<ListPartsElement> listParts) {
-
-}
