--- publication-rest/src/main/java/no/unit/nva/publication/uploadfile/restmodel/PrepareUploadPartResponseBody.java
@@ -1,7 +1,0 @@
-package no.unit.nva.publication.uploadfile.restmodel;
-
-import java.net.URL;
-
-public record PrepareUploadPartResponseBody(URL url) {
-
-}
